css cleanup. basic support for shopping list dates.
removed all render_to_response (which doesn't pass the request context (ie the csrf token) and is soon to be deprecated). replaced with render.

added generateList view which grabs the start and end date and for now, prints them out.

removed month.css and moved the td default style to a calendar style.

added a jquery datepicker.